detecting IPSEC SPI unchanged
This commit introduce a control during the `LoadIPSecKeys` to return
an error when providing a new key (e.g., during rotation) without
incrementing the SPI. In this fix, such case is intercepted and logged
with an error, while preserving status and policies referring to the
latest/previous valid key by not inserting the key in `ipSecKeysRemovalTime`.
Moreover, this commit adjusts the SPI fields in the test data so that they
can preserve their expected behaviour. Finally, a new test is introduced
to verify that `LoadIPSecKeys` effectively returns an error when the SPI
does not while loading a new IPSec key.

Expected error in the daemons logs while reproducing after the fix:
`invalid SPI: changing IPSec keys requires incrementing the key id`.

Fixes: #14873

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>