fix add spi=0 to ipSecKeysRemovalTime
This commit prevents the `LoadIPSecKeys` function to add to the
`ipSecKeysRemovalTime` map the value SPI 0. To the best of my understanding,
we don't allow to provide a key with spi=0, which will cause cilium to
return an error with the proper "invalid spi value" message. At this point,
the 1st time that an IPsec key is loaded, this function performs an
unneeded update of the map with the key=0 and value=time.Now(), which will
never be removed.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>