refactoring tracing in `do_netdev` to add src identity if available
This commit slightly refactors the tracing of the packets in the
`do_netdev` function. Before this commit, we always trace the packet
with an unknown src identity. Since few lines of code later we retrieve
this value, it could make sense to use it also while tracing the packet
so that we provide a bit more context. An example of output after applying
this patch would be as follows (no encryption, native routing):

```bash
1:<- endpoint 2223 flow 0x0 , identity 2173->unknown state unknown ifindex 0 orig-ip 0.0.0.0: 10.244.3.213 -> 10.244.1.98 EchoRequest
2:-> stack flow 0x0 , identity 2173->13606 state new ifindex 0 orig-ip 0.0.0.0: 10.244.3.213 -> 10.244.1.98 EchoRequest
3:-> 0: 10.244.3.213 -> 10.244.1.98 EchoRequest
4:<- network flow 0x0 , identity 13606->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: 10.244.1.98 -> 10.244.3.213 EchoReply
5:<- host flow 0x0 , identity 13606->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: 10.244.1.98 -> 10.244.3.213 EchoReply
6:-> endpoint 2223 flow 0x0 , identity 13606->2173 state reply ifindex lxcf91eb0f2f0ab orig-ip 10.244.1.98: 10.244.1.98 -> 10.244.3.213 EchoReply
```

We can now see in line 4 the source identity while logging the packet.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>