wireguard tracing ingress traffic with TRACE_REASON_ENCRYPTED
This commit adds logic to trace ingress WireGuard packets with the
TRACE_REASON_ENCRYPT. The reason is to have an aligned behavior similarly
to IPSec, for which we inspect and check whether the packet is ESP and mark
it for decryption. Within WireGuard, we just need to check that it is a
valid WireGuard packet for and from a cilium-managed endpoint and trace it
accordingly without having to use ctx->mark. An example output would look
as follows:

```bash
1:<- endpoint 321 flow 0x0 , identity 41978->unknown state unknown ifindex 0 orig-ip 0.0.0.0: 10.244.2.90 -> 10.244.1.17 EchoRequest
2:-> stack flow 0x0 , identity 41978->26105 state new ifindex 0 orig-ip 0.0.0.0: 10.244.2.90 -> 10.244.1.17 EchoRequest
3:-> 0: 10.244.2.90 -> 10.244.1.17 EchoRequest
4--> network flow 0x0 , identity unknown->unknown state new ifindex eth0 orig-ip 0.0.0.0: 172.18.0.4:51871 -> 172.18.0.2:51871 udp
5-<- network encrypted  flow 0x0 , identity remote-node->unknown state new ifindex eth0 orig-ip 0.0.0.0: 172.18.0.2:51871 -> 172.18.0.4:51871 udp
6:<- host flow 0x0 , identity 26105->unknown state unknown ifindex cilium_wg0 orig-ip 0.0.0.0: 10.244.1.17 -> 10.244.2.90 EchoReply
7:-> endpoint 321 flow 0x0 , identity 26105->41978 state reply ifindex lxceef65ccc7619 orig-ip 10.244.1.17: 10.244.1.17 -> 10.244.2.90 EchoReply
```

Note how in line 5 the FROM_NETWORK packet is marked as `encrypted`.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>