bpf: pull out and refactor TRACE_REASON_ENCRYPTED from trace reason
This commit takes the TRACE_REASON_ENCRYPTED away from the trace.reason
in favor of the new classifiers added in previous commits.
The idea is to simplify programs and interpretability, leaving trace
reason for the actual reasons, and move flags into a separate variable.
While doing so, we introduce two new flags CLS_FLAG_IPSEC and CLS_FLAG_WIREGUARD.
This way we provide more fine-grained info to monitor/hubble for enhancing tracing.
The `IsEncrypted()` check now looks whether either one of the two flag
is set, but the functional behavior is preserved.

In addition, this also adds the IsEncrypted function to the DropNotify
events as we have for TraceNotify to align the behaviour.

This patch adds some extra info to our monitor output making use of the
IsIPSec()/IsWireguard() flags added in the previous patch.
With this, we can now tell the user which encryption is applied to the
traced packet.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>