bpf: introduce classifiers for WireGuard
This commits move the current logic in from_netdev to classify a network
packet into a unique function for clarity and testing purposes.
Let's add bpf tests to these new set of functions and remove the old
`wireguard_helpers_test.c`.

While refactoring the logic into the ctx_classify_by_pkt_hdr function,
we drop some requirement to classify WireGuard traffic. Prior to this
commit, we used to additionally verify that (a) both source and destination
L4 ports were equal to WG_PORT, and (b) that the identity was inside the
cluster. Let's now just check whether source or destination port is the
WireGuard in-use port.

In addition, as for bpf_host, let's use `ctx_classify_by_pkt_hdr{4,6}` also
in the from-wireguard hook. For now, it will not be set to any value,
but in subsequent commits we'll start marking overlay traffic.

Finally, let's rename the helper function ctx_mark_is_wireguard() to
ctx_is_wireguard() for clarity, to align the naming convention to
ctx_is_overlay() that we already had.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>