bpf: prepare for classifiers in trace/drop notification paths
This commit is a prep work for the upcoming commits. We refactor the trace
flags currently in use (ipv6, l3_dev) into an apposite file and data type
`cls_flags_t` now defined as a `__u8`. We noticed that, while handling
trace and drop notification from the userspace in monitor/hubble, the
datapath need to report common flags for both trace/drop notification.
Therefore, we here refactor that into a dedicated file with shared methods:

1. ctx_classify_by_eth_hlen(ctx):  base function to report whether the packet
                              is from an L3 device (CLS_FLAG_L3_DEV), and
                              in case set CLS_FLAG_IPV6 if ctx->protocol
                              is ETH_P_IPV6. This is done by looking at
                              the ETH_HLEN value.
2. ctx_classify_by_eth_hlen4(ctx): same as (1), added for completeness and
                              future extensions.
3. ctx_classify_by_eth_hlen6(ctx): same as (1) but always sets CLS_FLAG_IPV6.

In `struct trace_notify` and `struct drop_notify`, we leave the flags as
a `__u8` fields. This is useful for when in future the cls_flags_t data type
grows from __u8 into, for instance, __u16 because we need to handle more
and more flags. In that case, the compiler would report an error, and we
will know that we need to hack the implementation in the structures (ex.
split flags into multiple fields, as flags_lower/flags_upper). Trace and
Drop structures can grow independently, therefore they have to handle
this changes on their own, and cannot rely on a cls_flags_t field.

In addition, the commit adds the possibility to pass additional classifiers
to trace/drop notification events. While doing that, it tries to introduce
changes to all possible codepaths in which we might have possible flags
(upcoming commits). This is useful when we compute additional
flags with respect to the base ones, such as in from_netdev where we look
at L4 ports to match WireGuard packets.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>