bpf: adopt TRACE_PAYLOAD_LEN_OVERLAY transparently for overlay packets
This commit introduces a new variable alongside the existing
TRACE_PAYLOAD_LEN. The new TRACE_PAYLOAD_LEN_OVERLAY is supposed to
be used whenever the dataplane is able to classify a packet as OVERLAY.
This is to fix cases in which Hubble/Monitor fail to decode packet due to
`TCP data offset greater than packet length` error. This error can happen
with an TCPv6 SYN (options MSS) over VXLAN: we're pushing 128B, but 144B
would be needed for correctly decoding the TCP options.

The utility function `ctx_default_monitor_from_classifiers()`
returns the correct PAYLOAD_LEN by looking at the classifiers flags: if it
detects an overlay packet (VXLAN/Geneve), then TRACE_PAYLOAD_LEN_OVERLAY
is returned, otherwise the old TRACE_PAYLOAD_LEN.

For drop notify events, the returned value is always used: we previously
used TRACE_PAYLOAD_LEN no matter what the packet was, so let's patch and
push those additional bytes for overlay packets.

For trace notify events, we also account for the `monitor` value provided:

* if `monitor` is 0, then the utility function is called, and the returned
  value used accordingly. Since `monitor` is not updated, the event is
  emitted (or not) according to current aggregation level.
* if monitor is TRACE_PAYLOAD_LEN, then we still call the utility function
  to decide how many bytes to push. Given that so far TRACE_PAYLOAD_LEN
  was the default value for those cases in which `monitor` is not provided
  by the ct lookups, we now want to account also for the additional bytes
  needed to decode overlay IPv6 traffic. To do so, we could either (a) modify
  the monitor value in all the occurrences after computing classifiers in
  the code, or (b) do it once as this patch proposes before emitting the
  event.

In brief, let's keep using TRACE_PAYLOAD_LEN as the default value to be
used in the code, but adjust it transparently to TRACE_PAYLOAD_LEN_OVERLAY
when tracing an overlay packet.

With this patch, the Monitor is now able to decode such packets (previously
failing with "[error]" and the hidden error message was, for instance,
"TCP data offset greater than packet length"):

```bash
<- network flow 0x30bee5c6 , identity remote-node->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: [fd00:10:244:1::99a]:8080 -> [fd00:10:244:2::1cc9]:34370 tcp SYN [tunnel 172.18.0.4:56309 -> 172.18.0.3:8472 vxlan]
<- network flow 0x30bee5c6 , identity remote-node->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: [fd00:10:244:1::99a]:8080 -> [fd00:10:244:2::1cc9]:34370 tcp ACK [tunnel 172.18.0.4:56309 -> 172.18.0.3:8472 vxlan]
<- network flow 0x30bee5c6 , identity remote-node->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: [fd00:10:244:1::99a]:8080 -> [fd00:10:244:2::1cc9]:34370 tcp ACK [tunnel 172.18.0.4:56309 -> 172.18.0.3:8472 vxlan]
<- network flow 0x30bee5c6 , identity remote-node->unknown state unknown ifindex eth0 orig-ip 0.0.0.0: [fd00:10:244:1::99a]:8080 -> [fd00:10:244:2::1cc9]:34370 tcp ACK, FIN [tunnel 172.18.0.4:56309 -> 172.18.0.3:8472 vxlan]
```

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>